Added Thread Policy

+++ app/Http/Controllers/ThreadsController.php
@@ -121,6 +121,15 @@ class ThreadsController extends Controller
     public function destroy($channel, Thread $thread)
     {
         //
+        // if($thread->user_id != auth()->id()){
+            // if(request()->wantsJson()){
+            //     return response(['status' => 'Permission Denied'], 403);
+            // }
+
+            // return redirect('/login');
+        //     abort('403','You do not have permission to do this');
+        // }
+            $this->authorize('update', $thread);
             $thread->replies()->delete();
             $thread->delete();
             if(request()->wantsJson){return response ([], 204);}
